Added function to load certs from the system
Windows doesn't respect the openssl standard, so that needed to be handled manuall.

Since this project uses a statically linked openssl, the default locations may be inaccurate. If that directory doesn't exist, it tries to get it from the system openssl via a popen. If that also fails, it will not be able to load system certs.